requirements.txt
fastapi
uvicorn
pystan==2.19.1.1
setuptools
prophet
u8darts[prophet]